Added glossary terms tab
