+++ openmetadata-ui/src/main/resources/ui/src/interface/types.d.ts
@@ -104,6 +104,7 @@ declare module 'Models' {
   };
 
   export interface AssetsDataType {
+    isLoading?: boolean;
     data: FormattedTableData[];
     total: number;
     currPage: number;
