+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/en-us.json
@@ -239,6 +239,7 @@
     "generate": "Generate",
     "glossary": "Glossary",
     "glossary-term": "Glossary Term",
+    "glossary-term-plural": "Glossary Terms",
     "go-back": "Go Back",
     "google": "Google",
     "govern": "Govern",
