+++ openmetadata-ui/src/main/resources/ui/src/rest/glossaryAPI.ts
@@ -15,7 +15,6 @@ import { AxiosResponse } from 'axios';
 import { Operation } from 'fast-json-patch';
 import { Include } from 'generated/type/include';
 import { PagingResponse } from 'Models';
-import { ModifiedGlossaryData } from 'pages/Glossary/GlossaryPageV1.component';
 import { CreateGlossary } from '../generated/api/data/createGlossary';
 import { CreateGlossaryTerm } from '../generated/api/data/createGlossaryTerm';
 import { Glossary } from '../generated/entity/data/glossary';
@@ -31,22 +30,12 @@ type Params = {
   include?: Include;
 };
 
-const BASE_URL = '/glossaries';
-
-export const getGlossaries = async (
-  paging = '',
-  limit = 10,
-  arrQueryFields: string | string[] = ''
-) => {
-  const qParams = `limit=${limit}`;
-  const url = getURLWithQueryFields(`/glossaries`, arrQueryFields, qParams);
-
-  const response = await APIClient.get<{ data: ModifiedGlossaryData[] }>(
-    paging ? `${url}&${paging}` : url
-  );
-
-  return response.data;
+export type ListGlossaryTermsParams = Params & {
+  glossary?: string;
+  parent?: string;
 };
+
+const BASE_URL = '/glossaries';
 
 export const getGlossariesList = async (params?: Params) => {
   const response = await APIClient.get<PagingResponse<Glossary[]>>(BASE_URL, {
@@ -103,16 +92,15 @@ export const getGlossariesByName = async (
   return response.data;
 };
 
-export const getGlossaryTerms = (
-  glossaryId = '',
-  limit = 10,
-  arrQueryFields: string | string[] = ''
-): Promise<AxiosResponse<{ data: GlossaryTerm[] }>> => {
-  let qParams = `limit=${limit}`;
-  qParams += glossaryId ? `&glossary=${glossaryId}` : '';
-  const url = getURLWithQueryFields(`/glossaryTerms`, arrQueryFields, qParams);
+export const getGlossaryTerms = async (params: ListGlossaryTermsParams) => {
+  const response = await APIClient.get<PagingResponse<GlossaryTerm[]>>(
+    '/glossaryTerms',
+    {
+      params,
+    }
+  );
 
-  return APIClient.get(url);
+  return response.data;
 };
 
 export const getGlossaryTermsById = (
