+++ openmetadata-ui/src/main/resources/ui/src/styles/spacing.less
@@ -467,7 +467,10 @@
 .w-80 {
   width: 80%;
 }
-
+.px-1\.5 {
+  padding-left: 0.375rem /* 6px */;
+  padding-right: 0.375rem /* 6px */;
+}
 .px-8 {
   padding-left: 2rem /* 32px */;
   padding-right: 2rem /* 32px */;
