+++ openmetadata-ui/src/main/resources/ui/src/utils/GlossaryUtils.ts
@@ -12,29 +12,14 @@
  */
 
 import { AxiosError } from 'axios';
-import { t } from 'i18next';
-import { cloneDeep, isEmpty } from 'lodash';
-import { ModifiedGlossaryData } from 'pages/Glossary/GlossaryPageV1.component';
-import { DataNode } from 'rc-tree/lib/interface';
-import {
-  getGlossaries,
-  getGlossaryTermByFQN,
-  getGlossaryTerms,
-} from 'rest/glossaryAPI';
 import { searchData } from 'rest/miscAPI';
-import {
-  FQN_SEPARATOR_CHAR,
-  WILD_CARD_CHAR,
-} from '../constants/char.constants';
-import { PRIMERY_COLOR, TEXT_BODY_COLOR } from '../constants/constants';
+import { WILD_CARD_CHAR } from '../constants/char.constants';
 import { SearchIndex } from '../enums/search.enum';
 import { Glossary } from '../generated/entity/data/glossary';
 import { GlossaryTerm } from '../generated/entity/data/glossaryTerm';
 import { EntityReference } from '../generated/type/entityReference';
 import { SearchResponse } from '../interface/search.interface';
-import { FileIcon, FolderIcon } from '../utils/svgconstant';
 import { formatSearchGlossaryTermResponse } from './APIUtils';
-import { getEntityName } from './CommonUtils';
 
 export interface GlossaryTermTreeNode {
   children?: GlossaryTermTreeNode[];
@@ -68,324 +53,6 @@ export const getGlossaryTermlist = (
   terms: Array<GlossaryTerm> = []
 ): Array<string> => {
   return terms.map((term: GlossaryTerm) => term.fullyQualifiedName || '');
-};
-
-/**
- * To get child terms of any node if available
- * @param listTermFQN fqn of targeted child terms
- * @returns promise of list of glossary terms
- */
-export const getChildGlossaryTerms = (listTermFQN: Array<string>) => {
-  return new Promise<GlossaryTerm[]>((resolve, reject) => {
-    const promises = listTermFQN.map((term) => {
-      return getGlossaryTermByFQN(term, ['children']);
-    });
-    Promise.allSettled(promises)
-      .then((responses) => {
-        const data = responses.reduce((prev, curr) => {
-          return curr.status === 'fulfilled' ? [...prev, curr.value] : prev;
-        }, [] as GlossaryTerm[]);
-        resolve(data);
-      })
-      .catch((err) => {
-        reject(err);
-      });
-  });
-};
-
-/**
- * To recursively generate RcTree data from glossary list
- * @param data list of glossary or glossary terms
- * @returns RcTree data node
- */
-export const generateTreeData = (data: ModifiedGlossaryData[]): DataNode[] => {
-  return data.map((d) => {
-    return d.children?.length
-      ? {
-          key: (d as GlossaryTerm)?.fullyQualifiedName || d.name,
-          title: getEntityName(d),
-          children: generateTreeData(d.children as ModifiedGlossaryData[]),
-          data: d,
-          icon: ({ selected }) =>
-            FolderIcon(selected ? PRIMERY_COLOR : TEXT_BODY_COLOR),
-        }
-      : {
-          key: (d as GlossaryTerm)?.fullyQualifiedName || d.name,
-          title: getEntityName(d),
-          data: d,
-          icon: ({ selected }) =>
-            FileIcon(selected ? PRIMERY_COLOR : TEXT_BODY_COLOR),
-        };
-  });
-};
-
-/**
- * Creates glossary term tree node from fqn
- * and root node name
- * @param leafFqn node fqn
- * @param name root node name
- * @returns node for glossary tree
- */
-const createGlossaryTermNode = (
-  leafFqn: string,
-  name: string
-): GlossaryTermTreeNode => {
-  const arrFQN = leafFqn.split(`${name}${FQN_SEPARATOR_CHAR}`);
-  const childName = arrFQN[1]?.split(FQN_SEPARATOR_CHAR)[0];
-
-  return !childName
-    ? {
-        name,
-        fullyQualifiedName: arrFQN[0],
-      }
-    : {
-        name,
-        fullyQualifiedName: `${arrFQN[0]}${name}`,
-        children: [createGlossaryTermNode(leafFqn, childName)],
-      };
-};
-
-/**
- * To merge the duplicate glossaries and terms
- * to generate optimised tree
- * @param treeNodes list of glossary nodes with duplicate items
- * @returns list of glossary nodes with unique items
- */
-const optimiseGlossaryTermTree = (treeNodes?: GlossaryTermTreeNode[]) => {
-  if (treeNodes) {
-    for (let i = 0; i < treeNodes.length; i++) {
-      // const reps = 0;
-      for (let j = i + 1; j < treeNodes.length; ) {
-        if (
-          treeNodes[j].fullyQualifiedName === treeNodes[i].fullyQualifiedName
-        ) {
-          if (treeNodes[j].children?.length || treeNodes[i].children?.length) {
-            treeNodes[i].children = (treeNodes[i].children || []).concat(
-              treeNodes[j].children || []
-            );
-          }
-          treeNodes.splice(j, 1);
-        } else {
-          j++;
-        }
-      }
-      if (treeNodes[i].children?.length) {
-        treeNodes[i].children = optimiseGlossaryTermTree(treeNodes[i].children);
-      }
-    }
-  }
-
-  return treeNodes;
-};
-
-/**
- * To generate glossry tree from searched terms
- * @param searchedTerms list of formatted searched terms
- * @returns list of glossary tree
- */
-export const getSearchedGlossaryTermTree = (
-  searchedTerms: GlossaryTerm[]
-): GlossaryTermTreeNode[] => {
-  const termTree: GlossaryTermTreeNode[] = [];
-  for (const term of searchedTerms) {
-    if (term.fullyQualifiedName) {
-      const arrFQN = term.fullyQualifiedName.split(FQN_SEPARATOR_CHAR);
-      const rootName = arrFQN[0];
-      termTree.push(createGlossaryTermNode(term.fullyQualifiedName, rootName));
-    }
-  }
-  optimiseGlossaryTermTree(termTree);
-
-  return termTree;
-};
-
-/**
- * To get Tree of glossaries based on search result
- * @param glossaries list of glossaries
- * @param searchedTerms list of formatted searched terms
- * @returns glossary list based on searched terms
- */
-export const updateGlossaryListBySearchedTerms = (
-  glossaries: ModifiedGlossaryData[],
-  searchedTerms: GlossaryTerm[]
-) => {
-  const searchedTermTree = getSearchedGlossaryTermTree(searchedTerms);
-
-  return glossaries.reduce((prev, curr) => {
-    const children = searchedTermTree.find(
-      (item) => item.name === curr.name
-    )?.children;
-    const data = (
-      children ? { ...curr, children: children } : curr
-    ) as ModifiedGlossaryData;
-
-    return [...prev, data];
-  }, [] as ModifiedGlossaryData[]);
-};
-
-/**
- * To get actions for action dropdown button
- * @returns list of action items
- */
-export const getActionsList = () => {
-  return [
-    {
-      name: t('label.add-entity', { entity: t('label.term') }),
-      value: 'add_term',
-    },
-  ];
-};
-
-/**
- * To get hierarchy of fqns from glossary to targeted term
- * from given fqn
- * @param fqn fqn of glossary or glossary term
- * @returns list of fqns
- */
-export const getHierarchicalKeysByFQN = (fqn?: string) => {
-  if (fqn) {
-    const keys = fqn.split(FQN_SEPARATOR_CHAR).reduce((prev, curr) => {
-      const currFqn = prev.length
-        ? `${prev[prev.length - 1]}${FQN_SEPARATOR_CHAR}${curr}`
-        : curr;
-
-      return [...prev, currFqn];
-    }, [] as string[]);
-
-    return keys;
-  } else {
-    return [];
-  }
-};
-
-/**
- * To get glossary term data from glossary object
- * @param glossary parent glossary
- * @param termFqn fqn of targeted glossary term
- * @returns Glossary term or {}
- */
-export const getTermDataFromGlossary = (
-  glossary: ModifiedGlossaryData,
-  termFqn: string
-) => {
-  let data: ModifiedGlossaryData | GlossaryTerm = cloneDeep(glossary);
-  const arrFQN = getHierarchicalKeysByFQN(termFqn);
-  for (let i = 1; i < arrFQN.length; i++) {
-    data = data?.children
-      ? ((data.children as unknown as GlossaryTerm[])?.find(
-          (item) =>
-            item.fullyQualifiedName === arrFQN[i] || item.name === arrFQN[i]
-        ) as GlossaryTerm)
-      : ({} as GlossaryTerm);
-    if (isEmpty(data)) {
-      break;
-    }
-  }
-
-  return data;
-};
-
-/**
- * To get relative indexed position of
- * glossary term from tree of glossaries
- * @param arrGlossary list of glossary
- * @param termFqn fqn of target glossary term
- * @returns array of numbered positions
- */
-export const getTermPosFromGlossaries = (
-  arrGlossary: ModifiedGlossaryData[],
-  termFqn: string
-) => {
-  const arrFQN = getHierarchicalKeysByFQN(termFqn);
-  const glossaryIdx = arrGlossary.findIndex((item) => item.name === arrFQN[0]);
-  const pos = [];
-  if (glossaryIdx !== -1) {
-    pos.push(glossaryIdx);
-    let data: ModifiedGlossaryData | GlossaryTerm = arrGlossary[glossaryIdx];
-    for (let i = 1; i < arrFQN.length; i++) {
-      const index = data?.children
-        ? (data.children as unknown as GlossaryTerm[])?.findIndex(
-            (item) =>
-              item.fullyQualifiedName === arrFQN[i] || item.name === arrFQN[i]
-          )
-        : -1;
-
-      if (index === -1) {
-        break;
-      }
-      data = (data?.children ? data?.children[index] : {}) as GlossaryTerm;
-      pos.push(index);
-    }
-  }
-
-  return pos;
-};
-
-/**
- * Fetches and adds root terms to each glossary
- * @param glossaries list of glossaries
- * @returns promise of list of glossaries with root terms
- */
-const getRootTermEmbeddedGlossary = (
-  glossaries: Array<ModifiedGlossaryData>
-): Promise<Array<ModifiedGlossaryData>> => {
-  return new Promise<Array<ModifiedGlossaryData>>((resolve, reject) => {
-    const promises = glossaries.map((glossary) =>
-      getGlossaryTerms(glossary.id, 1000, [
-        'children',
-        'relatedTerms',
-        'reviewers',
-        'tags',
-      ])
-    );
-    Promise.allSettled(promises)
-      .then((responses) => {
-        for (let i = 0; i < responses.length; i++) {
-          const res = responses[i];
-          const glossary = glossaries[i];
-          if (res.status === 'fulfilled') {
-            const { data } = res.value.data;
-
-            // Filtering root level terms from glossary
-            glossary.children = data.filter((d: GlossaryTerm) => !d.parent);
-          }
-        }
-        resolve(glossaries);
-      })
-      .catch((err: AxiosError) => {
-        reject(err);
-      });
-  });
-};
-
-/**
- * Fetches list of glossaries with root terms in each of them
- * @param paging pagination cursor
- * @param limit result count
- * @param arrQueryFields api query-string
- * @returns promise of api response
- */
-export const getGlossariesWithRootTerms = (
-  paging = '',
-  limit = 10,
-  arrQueryFields: string[]
-): Promise<Array<ModifiedGlossaryData>> => {
-  return new Promise<Array<ModifiedGlossaryData>>((resolve, reject) => {
-    getGlossaries(paging, limit, arrQueryFields)
-      .then((res) => {
-        const { data } = res;
-        if (data?.length) {
-          getRootTermEmbeddedGlossary(data)
-            .then((res) => resolve(res))
-            .catch((err) => reject(err));
-        } else {
-          resolve([]);
-        }
-      })
-      .catch((err: AxiosError) => {
-        reject(err);
-      });
-  });
 };
 
 export const getEntityReferenceFromGlossary = (
